refs #2 Add Follow notification
